Byttet ut AddFile med AddFileButton
