+++ src/components/oppgaver/DokumentasjonEtterspurtView.tsx
@@ -2,6 +2,7 @@ import React, {useState} from "react";
 import {Normaltekst} from "nav-frontend-typografi";
 import {
     DokumentasjonEtterspurt,
+    DokumentasjonEtterspurtElement,
     Fil,
     InnsynsdataActionTypeKeys,
     InnsynsdataSti,
@@ -19,7 +20,12 @@ import {REST_STATUS} from "../../utils/restUtils";
 import {SkjemaelementFeilmelding} from "nav-frontend-skjema";
 import DokumentasjonEtterspurtElementView from "./DokumentasjonEtterspurtElementView";
 import {onSendVedleggClicked} from "./onSendVedleggClicked";
-import {setFileUploadFailedVirusCheckInBackend} from "../../redux/innsynsdata/innsynsDataActions";
+import {
+    setFileUploadFailed,
+    setFileUploadFailedInBackend,
+    setFileUploadFailedVirusCheckInBackend,
+} from "../../redux/innsynsdata/innsynsDataActions";
+import {logInfoMessage} from "../../redux/innsynsdata/loggActions";
 
 interface Props {
     dokumentasjonEtterspurt: DokumentasjonEtterspurt;
@@ -64,7 +70,33 @@ const DokumentasjonEtterspurtView: React.FC<Props> = ({dokumentasjonEtterspurt, 
         listeOverDokumentasjonEtterspurtIderSomFeiletPaBackend.includes(dokumentasjonEtterspurt.oppgaveId) ||
         listeOverDokumentasjonEtterspurtIderSomFeiletIVirussjekkPaBackend.includes(dokumentasjonEtterspurt.oppgaveId);
 
-    const onAddFileChange = () => {};
+    const onAddFileChange = (
+        files: FileList,
+        internalIndex: number,
+        oppgaveElement: DokumentasjonEtterspurtElement
+    ) => {
+        dispatch(setFileUploadFailed(internalIndex.toString(), false));
+        dispatch(setFileUploadFailedInBackend(internalIndex.toString(), false));
+        dispatch(setFileUploadFailedVirusCheckInBackend(internalIndex.toString(), false));
+
+        Array.from(files).forEach((file: File) => {
+            if (!file) {
+                logInfoMessage("Tom fil ble forsøkt lagt til i OppgaveView.VelgFil.onChange()");
+            } else {
+                dispatch({
+                    type: InnsynsdataActionTypeKeys.LEGG_TIL_FIL_FOR_OPPLASTING,
+                    internalIndex: internalIndex,
+                    oppgaveElement: oppgaveElement,
+                    externalIndex: oppgaveIndex,
+                    fil: {
+                        filnavn: file.name,
+                        status: "INITIALISERT",
+                        file: file,
+                    },
+                });
+            }
+        });
+    };
 
     return (
         <div>
@@ -112,14 +144,14 @@ const DokumentasjonEtterspurtView: React.FC<Props> = ({dokumentasjonEtterspurt, 
                     return (
                         <DokumentasjonEtterspurtElementView
                             key={oppgaveElementIndex}
-                            typeTekst={typeTekst}
-                            tilleggsinfoTekst={tilleggsinfoTekst}
+                            tittel={typeTekst}
+                            beskrivelse={tilleggsinfoTekst}
                             oppgaveElement={oppgaveElement}
                             oppgaveElementIndex={oppgaveElementIndex}
-                            oppgaveIndex={oppgaveIndex}
                             oppgaveId={dokumentasjonEtterspurt.oppgaveId}
                             setOverMaksStorrelse={setOverMaksStorrelse}
                             onDelete={onDelete}
+                            onAddFileChange={onAddFileChange}
                         />
                     );
                 })}
